+++ src/Notifier/LaravelNotifier.php
@@ -2,7 +2,6 @@
 
 use Closure;
 use Illuminate\Mail\Mailer as Mail;
-use Illuminate\Support\Fluent;
 
 class LaravelNotifier implements NotifierInterface
 {
@@ -26,12 +25,12 @@ class LaravelNotifier implements NotifierInterface
     /**
      * Send notification via API.
      *
-     * @param  RecipientInterface           $user
-     * @param  \Illuminate\Support\Fluent   $message
-     * @param  \Closure                     $callback
+     * @param  RecipientInterface   $user
+     * @param  Message              $message
+     * @param  \Closure             $callback
      * @return Receipt
      */
-    public function send(RecipientInterface $user, Fluent $message, Closure $callback = null)
+    public function send(RecipientInterface $user, Message $message, Closure $callback = null)
     {
         $view    = $message->view;
         $data    = $message->data;