+++ src/Notifier/NotifierInterface.php
@@ -1,17 +1,16 @@
 <?php namespace Orchestra\Notifier;
 
 use Closure;
-use Illuminate\Support\Fluent;
 
 interface NotifierInterface
 {
     /**
      * Send notification via API.
      *
-     * @param  RecipientInterface           $user
-     * @param  \Illuminate\Support\Fluent   $message
-     * @param  \Closure                     $callback
+     * @param  RecipientInterface   $user
+     * @param  Message              $message
+     * @param  \Closure             $callback
      * @return Receipt
      */
-    public function send(RecipientInterface $user, Fluent $message, Closure $callback = null);
+    public function send(RecipientInterface $user, Message $message, Closure $callback = null);
 }