+++ src/Notifier/OrchestraNotifier.php
@@ -2,7 +2,6 @@
 
 use Closure;
 use Illuminate\Support\SerializableClosure;
-use Illuminate\Support\Fluent;
 use Orchestra\Memory\ContainerTrait;
 use Orchestra\Memory\Provider;
 
@@ -30,12 +29,12 @@ class OrchestraNotifier implements NotifierInterface
     /**
      * Send notification via API.
      *
-     * @param  RecipientInterface           $user
-     * @param  \Illuminate\Support\Fluent   $message
-     * @param  \Closure                     $callback
+     * @param  RecipientInterface   $user
+     * @param  Message              $message
+     * @param  \Closure             $callback
      * @return Receipt
      */
-    public function send(RecipientInterface $user, Fluent $message, Closure $callback = null)
+    public function send(RecipientInterface $user, Message $message, Closure $callback = null)
     {
         $view    = $message->view;
         $data    = $message->data;